pyproject.toml
[tool.poetry]
name = "cpr"
version = "1.0.0"
description = "A calibrated peer review Discord bot."
license = "MIT License"
authors = ["Mint <developer@chocomint.dev>"]
maintainers = ["Mint <developer@chocomint.dev>"]
readme = "README.md"
repository = "https://github.com/lunarmint/cpr-bot"

[tool.poetry.dependencies]
arrow = "~1.2.3"
asyncpraw = "~7.7.0"
coloredlogs = "~15.0.1"
confuse = "~2.0.0"
"discord.py" = "~2.2.2"
parsedatetime = "~2.6"
pymongo = "~4.3.3"
python = "^3.11"
python-magic = "~0.4.27"
requests-toolbelt = "~0.10.1"

[tool.poetry.dev-dependencies]
black = ">=23.1.0"
flake8 = ">=6.0.0"
python-magic-bin = ">=0.4.14" # For Windows development environment. Run "pip uninstall python-magic".

[tool.black]
line-length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"